feat: chakra ssr 설정 외 NavLink 분할
