+++ src/common/components/NavLinkMenu.tsx
@@ -1,50 +1,17 @@
-import { ReactNode } from 'react';
 import { CloseIcon, HamburgerIcon } from '@chakra-ui/icons';
 import {
 	Box,
 	Flex,
 	Icon,
 	IconButton,
-	Link,
 	Slide,
 	Stack,
 	StackDivider,
 	useColorModeValue,
 	useDisclosure,
 } from '@chakra-ui/react';
-import NextLink, { LinkProps } from 'next/link';
 import { AiFillGithub } from 'react-icons/ai';
-
-interface NavLinkProps extends Omit<LinkProps, 'href'> {
-	children: ReactNode;
-	to: string;
-}
-
-const NavLink = ({ children, to, ...props }: NavLinkProps) => {
-	const InitialColors = useColorModeValue('blackAlpha.700', 'whiteAlpha.700');
-	const hoverColors = useColorModeValue('blackAlpha.900', 'whiteAlpha.900');
-
-	return (
-		<NextLink {...props} href={to}>
-			<Link
-				px={2}
-				py={1}
-				rounded={'md'}
-				color={InitialColors}
-				_hover={{
-					textDecoration: 'none',
-					color: hoverColors,
-					transition: 'all .5s ease',
-					'& svg': {
-						color: hoverColors,
-					},
-				}}
-			>
-				{children}
-			</Link>
-		</NextLink>
-	);
-};
+import { NavLink } from './NavLink';
 
 export const NavLinkMenu = () => {
 	const { isOpen, onOpen, onClose } = useDisclosure();
