+++ src/pages/_app.tsx
@@ -1,10 +1,9 @@
-import { ChakraProvider } from '@chakra-ui/react';
 import { DefaultSeo } from 'next-seo';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 
 import { Layouts } from '@common/components/Layouts';
-import { customTheme } from '@core/theme';
+import { Chakra } from '@core/Chakra';
 import defaultSEOConfig from '../../next-seo.config';
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
@@ -17,11 +16,11 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
 				/>
 			</Head>
 			<DefaultSeo {...defaultSEOConfig} />
-			<ChakraProvider theme={customTheme}>
+			<Chakra cookies={pageProps.cookies}>
 				<Layouts>
 					<Component {...pageProps} />
 				</Layouts>
-			</ChakraProvider>
+			</Chakra>
 		</>
 	);
 };
