+++ src/pages/index.ts
@@ -1,3 +1,5 @@
 import { HomePage } from '@features/home/page/HomePage';
 
 export default HomePage;
+
+export { getServerSideProps } from '@core/Chakra';
